fix: api search-filter, api create post, add api search post by keyword
fix: api search-filter, api create post, add api search post by keyword

+++ src/main/java/com/n3c3/rentroom/dto/PostSearchDTO.java
@@ -5,8 +5,8 @@ import java.time.LocalDate;
 public class PostSearchDTO {
     private String title;
     private String address;
-    private Double minPrice;
-    private Double maxPrice;
+    private Long minPrice;
+    private Long maxPrice;
     private Double minRoomSize;
     private Double maxRoomSize;
     private String fullName;
@@ -32,19 +32,19 @@ public class PostSearchDTO {
         this.address = address;
     }
 
-    public Double getMinPrice() {
+    public Long getMinPrice() {
         return minPrice;
     }
 
-    public void setMinPrice(Double minPrice) {
+    public void setMinPrice(Long minPrice) {
         this.minPrice = minPrice;
     }
 
-    public Double getMaxPrice() {
+    public Long getMaxPrice() {
         return maxPrice;
     }
 
-    public void setMaxPrice(Double maxPrice) {
+    public void setMaxPrice(Long maxPrice) {
         this.maxPrice = maxPrice;
     }
